feat(티켓팅): 좌석 잠금, 예약 및 Toss 결제 흐름 구현 (프론트 연결전) fix(결제): JPA 좌석 조회, Redis 잠금 처리 및 Vite Toss SDK 설정 문제 해결

+++ spring-chat-matching/src/main/java/com/chatmatchingservice/springchatmatching/domain/event/dto/SectionResponseDto.java
@@ -9,6 +9,6 @@ public record SectionResponseDto(
         String grade,          // VIP / R / S
         int price,             // ticket.price
         int totalSeats,
-        int remainSeats,
+        Long remainSeats,
         List<SeatResponseDto> seats
 ) {}
